Added Modern Handgun asset pack
Modern Handgun asset pack has pre-made gun, bullets, and animation prefabs that are ready-to-use, just look for any assets that have "(Shooting)" in the name. I've done some editing to make it VR interactable, but it still needs a script for the bullet behavior. Also added a polyshape wall for testing in SampleScene.