Added Raycast to gun shooting
Made a separate "SampleSceneSam" for testing out stuff.
Added Raycast (laser beam) to SimpleShoot.cs
Note: there's a very slight recoil on the gun which causes it to tilt slightly and causing it to mess up the raycast. You'll probably want to lock the Rigidbody position/rotation (or set isKinematic=True), so it doesn't jiggle too much when the player shoots. Otherwise, kinda works most of the time (although it will sometimes not register).