Convert SampleSceneSam to Google Cardboard SDK
Converted SampleSceneSam to use the Google Cardboard SDK for the camara/controls. The VR Camera is messing up my ability to perceive lines. But the raycasting should shoot from the barrel of the gun to the center of the player's camera. If it doesn't look at SimpleShoot.cs and the raycasts.