Added TargetSpawner script and added Desert_Sam scene
Added Desert_Sam scene which is a copy of the WebXR samples and edited it for Google Cardboard. In the scene is a TargetSpawner object that spawns stuff on the left and moves to the right in a plane. On the target spawner is the TargetSpawner.cs script where you can adjust speed and target type in the inspector. Also fixed player camera not using skybox. Added more colors of targets
Added to README notes about VR optimization. Added invisible colliders to make sure bullets don't fly on forever.
TODO: more stuff to TargetSpawner (make sure it's on IgnoreRaycast Layer). Figure out how to bake shadows?